+++ src/index.ts
@@ -60,6 +60,7 @@ export {
   guessFreeChat,
   runsToString,
   toVideoId,
+  asString,
 } from "./utils";
 export * from "./yt";